requirements.txt
asgiref==3.8.1
asttokens==2.4.1
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
beautifulsoup4==4.12.3
beautify==0.1.4
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.7
decorator==5.1.1
Django==5.0.9
django-debug-toolbar==4.4.6
django-mssql-backend==2.8.1
mssql-django==1.5
django-plotly-dash==2.4.3
djangorestframework==3.15.2
dpd_components==0.1.0
et_xmlfile==2.0.0
executing==2.1.0
flake8==7.1.1
Flask==3.0.3
fonttools==4.55.2
fpdf==1.7.2
gunicorn
html2text==2024.2.26
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
mysqlclient==2.2.6
nest-asyncio==1.6.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.0
PyPDF2==3.0.1
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
pyxlsb==1.0.10
pyzmq==26.2.0
reportlab==4.2.5
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.0.6
whitenoise==6.8.2
zipp==3.21.0
zopfli==0.2.3.post1
